+++ src/Zen.Cli/Using.cs
@@ -1,0 +1,3 @@
+global using Zen.Core.SpectreConsole;
+global using Spectre.Console.Cli;
+global using Zen.Core.Constants;