+++ src/Zen.Core/Constants/DefaultUrls.cs
@@ -4,8 +4,9 @@ namespace Zen.Core.Constants
     {
         public const string IFCONFIG_URL = "http://ifconfig.me/all.json";
         public const string GITIGNORE_IO_URL = "https://www.gitignore.io";
-        public const string ANIME_CHAN_API = "https://animechan.vercel.app";
         public const string END_OF_LIFE_API = "https://endoflife.date";
+        public const string GENERATE_PASSWORD_API = "https://passwordinator.herokuapp.com/";
+        
         
         
     }