+++ src/Zen.Core/SpectreConsole/Integration/ZenCommandSetting.cs
@@ -1,0 +1,13 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Spectre.Console.Cli;
+
+namespace Zen.Core.SpectreConsole
+{
+    public class ZenCommandSetting : CommandSettings
+    {
+        
+    }
+}